♻️ Extracted data fetching methods from App.js to utils/data-fetchers.js
Reduced App.js from 675 to 461 lines (214 lines removed, 31.7% reduction).

Created utils/data-fetchers.js to contain:
- fetchApiData() - API calls with Ghost
- fetchDevData() - Dev/test mode data
- fetchOfferQueryStrData() - Offer preview parsing
- fetchQueryStrData() - Preview query string parsing
- fetchNotificationData() - Billing notification data
- fetchLinkData() - Portal link parsing
- fetchPreviewData() - Preview mode data
- fetchAllData() - Orchestrates all data sources

Benefits:
- Separated data fetching concerns from component logic
- Pure functions easier to test and reuse
- Reduced App.js complexity before hooks conversion
- All tests passing (256/257)
